Remove need for -sass and -js as dependencies #107
