+++ lib/cli.js
@@ -56,6 +56,7 @@ module.exports = function parseCLI(argv = process.argv.slice(2), help = HELP) {
 				"mode to avoid littering your file system with obsolete bundles");
 	}
 
+	// TODO: remove the baseuri option, use the setting from the manifest config instead
 	let { source, target, baseuri } = argv;
 	let rawConfig = require(path.resolve(process.cwd(), source));
 	let watchDirs = rawConfig.watchDirs || [];
@@ -63,18 +64,7 @@ module.exports = function parseCLI(argv = process.argv.slice(2), help = HELP) {
 	let config = {
 		aiur: [{
 			source,
-			target,
-			baseURI: baseuri
-		}],
-
-		sass: [{
-			source: "aiur/lib/style.scss",
-			target: `${target}/style-aiur.css`
-		}],
-
-		js: [{
-			source: "aiur/lib/client/index.js",
-			target: `${target}/script-aiur.js`
+			target
 		}],
 
 		watchDirs: [source, ...watchDirs],
@@ -85,13 +75,7 @@ module.exports = function parseCLI(argv = process.argv.slice(2), help = HELP) {
 			webRoot: target
 		},
 
-		plugins: [
-			{
-				bucket: "markup",
-				key: "aiur",
-				plugin: aiur
-			}
-		]
+		plugins: [aiur]
 	};
 
 	Object.entries(rawConfig.snippetAssets || {}).forEach(([pipeline, entries]) => {
