+++ lib/generate_layout.js
@@ -11,6 +11,8 @@ module.exports = async (navigation, manifest) => {
 		<meta name="viewport" content="width=device-width,initial-scale=1">
 		${page.meta}
 		<link href="${manifest.get("style-aiur.css")}" rel="stylesheet">
+		<link href="${manifest.get("prism.css")}" rel="stylesheet">
+		<script type="module" src="${manifest.get("script-aiur.js")}"></script>
 	</head>
 	<body id="aiur">
 		<nav id="aiur-nav">
@@ -21,7 +23,6 @@ module.exports = async (navigation, manifest) => {
 		<h1>${page.heading}${StatusBadge(page.status)}</h1>
 		${content}
 		</main>
-		<script src="${manifest.get("script-aiur.js")}"></script>
 	</body>
 </html>`;
 };
