+++ lib/index.js
@@ -1,7 +1,11 @@
 let Site = require("./site");
 
-module.exports = (config, assetManager, options) => {
+exports.key = "aiur";
+exports.bucket = "markup";
+
+exports.plugin = (config, assetManager, options) => {
 	let sites = config.map(siteConfig => new Site(siteConfig, assetManager));
+
 	return filepaths => {
 		let builds = sites.map(site => site.generate(filepaths));
 		return Promise.all(builds);
