+++ lib/site.js
@@ -1,8 +1,8 @@
 let generatePageTree = require("./tree");
-let { resolvePath } = require("faucet-pipeline-core/lib/util/resolve");
+let { resolvePath } = require("faucet-pipeline-core/lib/util");
 let { abort, reportFileStatus } = require("faucet-pipeline-core/lib/util");
 let { mkdirp } = require("mkdirp");
-let { writeFile } = require("node:fs/promises");
+let { readFile, writeFile } = require("node:fs/promises");
 let path = require("path");
 let PageRenderer = require("./page_renderer");
 let SnippetWrapper = require("./snippet_wrapper");
@@ -18,12 +18,12 @@ let DEFAULTS = {
 };
 
 module.exports = class Site {
-	constructor({ source, target, baseURI }, assetManager) {
+	constructor({ source, target }, assetManager) {
 		let { referenceDir } = assetManager;
 		Object.assign(this, {
 			source: resolvePath(source, referenceDir),
 			target: resolvePath(target, referenceDir, { enforceRelative: true }),
-			baseURI,
+			baseURI: assetManager.manifest.baseURI,
 			// TODO: either make this configurable or autoload renderers on use
 			renderers: DEFAULTS.renderers,
 			// TODO: make this configurable, but first find a simpler contract
@@ -34,6 +34,13 @@ module.exports = class Site {
 	}
 
 	async generate(filepaths) {
+		if(filepaths) {
+			// only copy the files on the first run
+			this.copyFile("prismjs/themes/prism.css", `${this.target}/prism.css`);
+			this.copyFile("aiur/lib/client/style.css", `${this.target}/style-aiur.css`);
+			this.copyFile("aiur/lib/client/index.js", `${this.target}/script-aiur.css`);
+		}
+
 		let tree = await this.tree();
 		// TODO: no longer necessary when we can use nite-owl 4.0.0
 		filepaths = new Set(filepaths);
@@ -106,6 +113,11 @@ module.exports = class Site {
 
 	get referenceDir() {
 		return this.assetManager.referenceDir;
+	}
+
+	async copyFile(source, destination) {
+		let content = await readFile(this.assetManager.resolvePath(source));
+		this.assetManager.writeFile(destination, content);
 	}
 };
