+++ lib/tree.js
@@ -1,5 +1,5 @@
 let Page = require("./page");
-let { resolvePath } = require("faucet-pipeline-core/lib/util/resolve");
+let { resolvePath } = require("faucet-pipeline-core/lib/util");
 
 module.exports = generatePageTree;
