+++ package.json
@@ -26,10 +26,7 @@
 	},
 	"dependencies": {
 		"commonmark": "^0.31.2",
-		"faucet-pipeline-core": "^3.0.0",
-		"faucet-pipeline-js": "^3.0.0",
-		"faucet-pipeline-sass": "^1.9.2",
-		"faucet-pipeline-static": "^2.0.0",
+		"faucet-pipeline-core": "git+https://github.com/faucet-pipeline/faucet-pipeline-core.git#util-cleanup",
 		"handlebars": "^4.7.7",
 		"metacolon": "^2.0.3",
 		"minimist": "^1.2.5",
@@ -38,7 +35,9 @@
 		"prismjs": "^1.26.0"
 	},
 	"devDependencies": {
+		"donny": "^0.3.0",
 		"eslint-config-fnd": "^1.12.0",
+		"nite-owl": "^5.0.5",
 		"release-util-fnd": "^3.0.0",
 		"rimraf": "^6.0.1"
 	}
